Changement du type de l'id

+++ Craft2Gether/src/main/java/fr/crafttogether/repositories/BlocRepository.java
@@ -3,5 +3,5 @@ package fr.crafttogether.repositories;
 import fr.crafttogether.models.Bloc;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BlocRepository extends JpaRepository<Bloc, Long> {
-        }
+public interface BlocRepository extends JpaRepository<Bloc, Integer> {
+}
